Refactor current parsing logic into a new SceneParser class.
This maintains Scene::load() for convenience, but separates the scene parser (responsible for creating objects based upon the content of the file) from the scene instance itself (responsible for managing objects).

This means that in future, the current XML parser could be one of several which process different scene formats, without needing to modify the way the scene itself works.

Conversely, in an extremely constrained system, it may be desirable to provide an implementation of a scene parser which doesn't have any external dependencies upon libxml2 or other libraries.