Create the first skeleton for a scene graph.
Scene is now a subclass of SceneObject. A SceneObject has, for the moment,
some debugging output enabled which shows when each object is created.

A parent internal member has been added to the SceneParser class which is
used while parsing to keep track of which SceneObject new objects should
be added to. At the start of a parsing operation, this will be set to the
Scene object that is being parsed into.