Add basic Autotools-based build machinery.
Because the project is now sufficiently complex that there are interdependencies between source files, dependency management (either automatic or manual) is a requirement in order to avoid mysterious linking and runtime errors.

Automake-based logic is the de-facto standard for cross-platform software, particularly that written in C and C++. This has the added advantage of allowing us to easily test for libxml2 as required.

Note that implementing Automake-based logic does not rule out generating project files necessary to build this project with an IDE such as Xcode or Visual Studio.