Move all of the XML-specifics into a private XMLSceneParser class.
This means that an application (such as scenetool) no longer needs to have
any knowledge or direct dependency upon libxml2 development files - although
the Yuka library will continue to depend upon it at runtime.

If support for other scene formats is added in the future, this insulation
will mean that it can be implemented near-transparently to any applications.