Add a simple reference-counted Yuka::Object base class.
This provides us with a standard pattern for object management which we can use for all kinds of classes - visible, non-visible, utility, and so on.

It also means there is now a common ancestor class to which utility methods can be added where needed.

Note that Yuka::Object::retain() and Yuka::Object::release() are not currently thread-safe, meaning they cannot safely be used for mutex-free ownership transfer between threads.