Promote generic functionality to traits.
Object tags are now part of Traits::Identifiable.
The ability to set() tags is now part of Traits::Scriptable.
The printing of tags as part of a property list is now part of Traits::Debuggable.

Note that these interactions are enabled through the use of RTTI, which may
have a performance penalty.

For Traits::Debuggable, this is not a concern -- debugging an object is not
expected to be performance-critical.

For Traits::Scriptable, a dynamic_cast<>() may prove too inefficient if, e.g.,
set() is invoked as part of an event handler which executes for each frame.