Make Traits and Behaviours work together.
This means that you can create a Transform behaviour, and attach it to any Flexible, Spatial object, and the spatial state will be updated automatically according to the transform properties.

In doing so, this change removed various bits of dead code, and instilled the pattern whereby Traits are always derived from using the virtual inheritance mechanism, so as to ensure that the properties of the core Trait class are never duplicated due to diamond-shaped inheritance structures.

Within the Debuggable trait, there is now specific handling for other traits, including a printTraits() method. Note that the default implementation uses the dynamic_cast<Traits::SomeTrait *>(this) pattern to test the availability of traits and access members.

Adding new traits to the engine is straightforward; unlike with new types of object and behaviours, adding new traits outside of the engine is more difficult to achieve feature parity with, because the Debuggable trait has compile-time knowledge of the other bundled traits, which would not apply to a user-defined trait. In any case, the nature of traits and behaviours is such that many use-cases of user-defined traits could probably be achieved with user-defined behaviours instead.